requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
yfinance>=0.2.18
matplotlib>=3.7
pyyaml>=6.0
quantstats>=0.0.62
vectorbt>=0.26.0
pytest>=7.0
